CLI-837 Consolidate version helpers into core/version.ts
