+++ tests/unit/commands/update/update-version.test.ts
@@ -64,11 +64,15 @@ void mock.module('@/core/host/platform-detector.js', () => ({
 }));
 
 // Mock the shared version helpers used by the CLI Version class.
-const { isNewerVersion: realIsNewerVersion, stripBuildNumber: realStripBuildNumber } =
-  await import('@/core/host/version.ts');
-void mock.module('@/core/host/version.ts', () => ({
+const {
+  isNewerVersion: realIsNewerVersion,
+  stripBuildNumber: realStripBuildNumber,
+  Version,
+} = await import('@/core/version.ts');
+void mock.module('@/core/version.ts', () => ({
   isNewerVersion: mock(realIsNewerVersion),
   stripBuildNumber: mock(realStripBuildNumber),
+  Version,
 }));
 
 const { checkForUpdate, fetchLatestVersion } = await import('@/commands/update/update-check.ts');
